Separate aml example test from pipeline
