+++ .azure_pipelines/olive-aml-ci.yaml
@@ -1,0 +1,47 @@
+trigger: none
+pr: none
+
+schedules:
+- cron: 0 6 * * 5
+  displayName: Scheduled Build
+  branches:
+    include:
+    - main
+  always: true
+
+jobs:
+# Linux examples test
+- template: job_templates/olive-example-template.yaml
+  parameters:
+    name: Linux_CI
+    pool: $(OLIVE_POOL_UBUNTU2004)
+    onnxruntime: onnxruntime
+    subfolder: azureml
+    examples:
+      bert_ptq_cpu_aml:
+        exampleFolder: bert
+        exampleName: bert_ptq_cpu_aml
+      resnet_ptq_cpu:
+        exampleFolder: resnet
+        exampleName: resnet_ptq_cpu_aml
+      resnet_vitis_ai_ptq_cpu:
+        exampleFolder: resnet
+        exampleName: resnet_vitis_ai_ptq_cpu_aml
+
+# Windows examples test
+- template: job_templates/olive-example-template.yaml
+  parameters:
+    name: Windows_CI
+    pool: $(OLIVE_POOL_WIN2019)
+    onnxruntime: onnxruntime
+    subfolder: azureml
+    examples:
+      bert_ptq_cpu_aml:
+        exampleFolder: bert
+        exampleName: bert_ptq_cpu_aml
+      resnet_ptq_cpu:
+        exampleFolder: resnet
+        exampleName: resnet_ptq_cpu_aml
+      resnet_vitis_ai_ptq_cpu:
+        exampleFolder: resnet
+        exampleName: resnet_vitis_ai_ptq_cpu_aml
