+++ .azure_pipelines/olive-ci.yaml
@@ -85,20 +85,14 @@ jobs:
     name: Linux_CI
     pool: $(OLIVE_POOL_UBUNTU2004)
     onnxruntime: onnxruntime
+    subfolder: local
     examples:
       bert_ptq_cpu:
         exampleFolder: bert
         exampleName: bert_ptq_cpu
-      bert_ptq_cpu_aml:
-        exampleFolder: bert
-        exampleName: bert_ptq_cpu_aml
       bert_ptq_cpu_docker:
         exampleFolder: bert
         exampleName: bert_ptq_cpu_docker
-      # running on linux is enough, only testing cuda ep on aml
-      bert_cuda_gpu_aml:
-        exampleFolder: bert
-        exampleName: bert_cuda_gpu_aml
       resnet_ptq_cpu:
         exampleFolder: resnet
         exampleName: resnet_ptq_cpu
@@ -127,13 +121,11 @@ jobs:
     name: Windows_CI
     pool: $(OLIVE_POOL_WIN2019)
     onnxruntime: onnxruntime
+    subfolder: local
     examples:
       bert_ptq_cpu:
         exampleFolder: bert
         exampleName: bert_ptq_cpu
-      bert_ptq_cpu_aml:
-        exampleFolder: bert
-        exampleName: bert_ptq_cpu_aml
       resnet_ptq_cpu:
         exampleFolder: resnet
         exampleName: resnet_ptq_cpu
@@ -156,6 +148,7 @@ jobs:
     name: Linux_GPU_CI
     pool: $(OLIVE_POOL_UBUNTU2004_GPU_V100)
     onnxruntime: onnxruntime-gpu
+    subfolder: local
     # torch 2.4.0 installs nvidia-cudnn-cu12 9.x but ort-gpu (with cuda 11.x) requires 8.x
     # pin torch version until os image has its own cudnn installed/updated which ort-gpu can use
     torch: torch==2.3.1
