+++ examples/test/azureml/test_resnet_ptq_cpu_aml.py
@@ -6,7 +6,9 @@ import os
 from pathlib import Path
 
 import pytest
-from utils import check_output, patch_config
+from onnxruntime import __version__ as OrtVersion
+from packaging import version
+from ..utils import check_output, patch_config
 
 from olive.common.utils import retry_func, run_subprocess
 
@@ -14,7 +16,7 @@ from olive.common.utils import retry_func, run_subprocess
 @pytest.fixture(scope="module", autouse=True)
 def setup():
     """Setups any state specific to the execution of the given module."""
-    cur_dir = Path(__file__).resolve().parent.parent
+    cur_dir = Path(__file__).resolve().parent.parent.parent
     example_dir = cur_dir / "resnet"
     os.chdir(example_dir)
 
@@ -28,9 +30,20 @@ def setup():
 
 @pytest.mark.parametrize("search_algorithm", ["random"])
 @pytest.mark.parametrize("execution_order", ["pass-by-pass"])
-@pytest.mark.parametrize("system", ["local_system"])
+@pytest.mark.parametrize("system", ["aml_system"])
 @pytest.mark.parametrize(
-    "olive_json", ["resnet_qat_default_train_loop_cpu.json", "resnet_qat_lightning_module_cpu.json"]
+    "olive_json",
+    [
+        "resnet_ptq_cpu.json",
+        # TODO(trajep): remove skip once the bug of azureml-fsspec is fixed
+        pytest.param(
+            "resnet_ptq_cpu_aml_dataset.json", marks=pytest.mark.skip(reason="credential bug in azureml-fsspec")
+        ),
+    ],
+)
+@pytest.mark.skipif(
+    version.parse(OrtVersion) == version.parse("1.16.0"),
+    reason="resnet is not supported in ORT 1.16.0 caused by https://github.com/microsoft/onnxruntime/issues/17627",
 )
 def test_resnet(search_algorithm, execution_order, system, olive_json):
     from olive.workflows import run as olive_run
