+++ examples/test/local/test_bert_cuda_gpu.py
@@ -6,30 +6,29 @@ import os
 from pathlib import Path
 
 import pytest
-from utils import check_output, patch_config
+from ..utils import check_output, patch_config
 
 
 @pytest.fixture(scope="module", autouse=True)
 def setup():
     """Setups any state specific to the execution of the given module."""
-    cur_dir = Path(__file__).resolve().parent.parent
+    cur_dir = Path(__file__).resolve().parent.parent.parent
     example_dir = cur_dir / "bert"
     os.chdir(example_dir)
     yield
     os.chdir(cur_dir)
 
 
-# NOTE: this doesn't need to run on a host with a GPU, as it's testing the CUDA EP on AML
 @pytest.mark.parametrize("search_algorithm", ["tpe"])
-@pytest.mark.parametrize("execution_order", ["joint"])
-@pytest.mark.parametrize("system", ["aml_system"])
+@pytest.mark.parametrize("execution_order", ["joint", "pass-by-pass"])
+@pytest.mark.parametrize("system", ["local_system"])
 @pytest.mark.parametrize("olive_json", ["bert_cuda_gpu.json"])
-def test_bert(search_algorithm, execution_order, system, olive_json):
+@pytest.mark.parametrize("enable_cuda_graph", [True, False])
+def test_bert(search_algorithm, execution_order, system, olive_json, enable_cuda_graph):
     from olive.workflows import run as olive_run
 
     olive_config = patch_config(olive_json, search_algorithm, execution_order, system, is_gpu=True)
-    # no need to run both pass_flows, only testing cuda ep on aml
-    del olive_config["pass_flows"]
+    olive_config["passes"]["perf_tuning"]["enable_cuda_graph"] = enable_cuda_graph
 
     footprint = olive_run(olive_config, tempdir=os.environ.get("OLIVE_TEMPDIR", None))
     check_output(footprint)
