+++ examples/test/local/test_bert_ptq_cpu.py
@@ -3,19 +3,16 @@
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
 import os
-import platform
 from pathlib import Path
 
 import pytest
-from utils import check_output, patch_config
-
-from olive.common.constants import OS
+from ..utils import check_output, patch_config
 
 
 @pytest.fixture(scope="module", autouse=True)
 def setup():
     """Setups any state specific to the execution of the given module."""
-    cur_dir = Path(__file__).resolve().parent.parent
+    cur_dir = Path(__file__).resolve().parent.parent.parent
     example_dir = cur_dir / "bert"
     os.chdir(example_dir)
     yield
@@ -24,12 +21,9 @@ def setup():
 
 @pytest.mark.parametrize("search_algorithm", ["tpe"])
 @pytest.mark.parametrize("execution_order", ["joint"])
-@pytest.mark.parametrize("system", ["docker_system"])
+@pytest.mark.parametrize("system", ["local_system"])
 @pytest.mark.parametrize("olive_json", ["bert_ptq_cpu.json"])
 def test_bert(search_algorithm, execution_order, system, olive_json):
-    if system == "docker_system" and platform.system() == OS.WINDOWS:
-        pytest.skip("Skip Linux containers on Windows host test case.")
-
     from olive.workflows import run as olive_run
 
     olive_config = patch_config(olive_json, search_algorithm, execution_order, system)
