+++ examples/test/local/test_mobilenet_qnn_ep.py
@@ -14,7 +14,7 @@ from olive.hardware import AcceleratorSpec
 @pytest.fixture(scope="module", autouse=True)
 def setup():
     """Setups any state specific to the execution of the given module."""
-    cur_dir = Path(__file__).resolve().parent.parent
+    cur_dir = Path(__file__).resolve().parent.parent.parent
     example_dir = str(cur_dir / "mobilenet")
     os.chdir(example_dir)
     sys.path.insert(0, example_dir)
