+++ examples/test/local/test_stable_diffusion_cuda_gpu.py
@@ -12,7 +12,7 @@ import pytest
 @pytest.fixture(scope="module", autouse=True)
 def setup():
     """Setups any state specific to the execution of the given module."""
-    cur_dir = Path(__file__).resolve().parent.parent
+    cur_dir = Path(__file__).resolve().parent.parent.parent
     example_dir = str(cur_dir / "stable_diffusion")
     os.chdir(example_dir)
     sys.path.insert(0, example_dir)
